TExecutorPool moved to a more suitable location
